LayoutTests/platform/ios-sim-deprecated/editing/secure-input/password-input-focusing-expected.txt

Verify that basic focusing/unfocusing updates secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state, no selection:
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused:
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

A regular text input is focused:
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused again:
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
 
